requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
asttokens==3.0.0
attrs==25.4.0
Authlib==1.6.5
av==13.1.0
backports.tarfile==1.2.0
beartype==0.22.5
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloup==3.0.8
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
cyclopts==4.2.3
debugpy==1.8.15
decorator==5.2.1
diskcache==5.6.3
Django==5.2.8
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.3
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.5
fastmcp==2.13.0.2
ffmpeg==1.4
gitdb==4.0.12
GitPython==3.1.45
glcontext==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isosurfaces==0.1.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==25.6.0
langchain==1.0.7
langchain-core==1.0.5
langchain-mcp-adapters==0.1.12
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.41
manim==0.19.0
ManimPango==0.6.1
mapbox_earcut==1.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mcp==1.20.0
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
more-itertools==10.8.0
narwhals==2.12.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
openapi-pydantic==0.5.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
parso==0.8.4
pathable==0.4.4
pathvalidate==3.3.1
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
pyglet==2.1.11
pyglm==2.8.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.28.0
rsa==4.9.1
scipy==1.16.3
screeninfo==0.8.1
six==1.17.0
skia-pathops==0.8.0.post2
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
srt==3.5.3
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.41.3
streamlit==1.51.0
svgelements==1.9.6
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.32.1
watchdog==6.0.0
wcwidth==0.2.13
websockets==14.2
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
